Add files via upload
I had some issues opening the main file but I still wrote the class. tell me what you guys think about having a child type enum in the parent class. I think it may make it easier to read what animal class we are dealing with if we are going through an array of animals.